tinkoff-app/requirements.txt
tinkoff-investments
mysql-connector-python
pandas
sqlalchemy
pymysql
